+++ src/createStatusBarItem.ts
@@ -4,22 +4,22 @@ import getStatusBarText from './helpers/getStatusBarText'
 
 // Taken from the `prettier-vscode` repo.
 const PRETTIER_STATUS_BAR_PRIORITY = -1
-const TOOLTIP_TEXT = 'Enable/Disable the formatter'
+const TOOLTIP_TEXT = 'Enable/Disable formatting'
 
 const createStatusBarItem = () => {
-  const statusBar = window.createStatusBarItem(
+  const statusBarItem = window.createStatusBarItem(
     StatusBarAlignment.Right,
     // We substract one so that it is always to the right of the Prettier status
     // bar item. If we didn’t do this, the position could switch randomly.
     PRETTIER_STATUS_BAR_PRIORITY - 1,
   )
 
-  statusBar.command = COMMAND_NAME
-  statusBar.tooltip = TOOLTIP_TEXT
-  statusBar.text = getStatusBarText()
-  statusBar.show()
+  statusBarItem.command = COMMAND_NAME
+  statusBarItem.tooltip = TOOLTIP_TEXT
+  statusBarItem.text = getStatusBarText()
+  statusBarItem.show()
 
-  return statusBar
+  return statusBarItem
 }
 
 export default createStatusBarItem