+++ src/extension.ts
@@ -1,15 +1,16 @@
-import { ExtensionContext } from 'vscode'
+import { ExtensionContext, workspace } from 'vscode'
 import registerCommand from './registerCommand'
 import createStatusBarItem from './createStatusBarItem'
-import getOnDidChangeConfigurationHandler from './getOnDidChangeConfigurationHandler'
+import getStatusBarText from './helpers/getStatusBarText'
 
 export function activate({ subscriptions }: ExtensionContext) {
   const command = registerCommand()
   const statusBarItem = createStatusBarItem()
-  const onDidChangeConfigurationHandler =
-    getOnDidChangeConfigurationHandler(statusBarItem)
+  const configurationChangeListener = workspace.onDidChangeConfiguration(() => {
+    statusBarItem.text = getStatusBarText()
+  })
 
   subscriptions.push(statusBarItem)
   subscriptions.push(command)
-  subscriptions.push(onDidChangeConfigurationHandler)
+  subscriptions.push(configurationChangeListener)
 }