+++ src/registerCommand.ts
@@ -1,32 +1,24 @@
-import { commands, ConfigurationTarget } from 'vscode'
-import {
-  COMMAND_NAME,
-  FORMATTING_SETTINGS,
-  DEFAULT_AFFECTS_CONFIGURATION,
-} from './constants'
-import getConfiguration from './helpers/getConfiguration'
+import { commands, ConfigurationTarget, workspace } from 'vscode'
+import { COMMAND_NAME, DEFAULT_AFFECTS_CONFIGURATION } from './constants'
 import getFormattingStatus from './helpers/getFormattingStatus'
 
 const registerCommand = () =>
   commands.registerCommand(COMMAND_NAME, () => {
-    const editorConfiguration = getConfiguration('editor')
-    const formattingToggleConfiguration = getConfiguration('formattingToggle')
-    const affectsConfiguration = formattingToggleConfiguration.get(
-      'affects',
+    const configuration = workspace.getConfiguration()
+    const affectsConfiguration = configuration.get(
+      'formattingToggle.affects',
       DEFAULT_AFFECTS_CONFIGURATION,
     )
     const isFormattingActivated = getFormattingStatus()
 
     // Updating the configuration will trigger the `onDidChangeConfiguration`
     // handler which will correctly update the text and icon in the status bar.
-    FORMATTING_SETTINGS.forEach(setting => {
-      if (affectsConfiguration.includes(setting)) {
-        editorConfiguration.update(
-          setting,
-          !isFormattingActivated,
-          ConfigurationTarget.Global,
-        )
-      }
+    affectsConfiguration.forEach(setting => {
+      configuration.update(
+        setting,
+        !isFormattingActivated,
+        ConfigurationTarget.Global,
+      )
     })
   })